Add files via upload
I updated my design to cater for an aduino Pro Micro which can behave like an HID when plugged in. So my current version does not need Mobiflight or vjoy. Just a proper joystick.ino to send the right outputs based on the steering inputs. I have updated the PCB also to enable Pro Micro and to enable daisy chaining by using a RJ10 connector for later extension. Find the PCB and ino details on github. Just note that the connectors from the Poti have now a different conbectivity - (Vcc - Arduino Pin - GND).